Update for style changes in python-wink
