+++ homeassistant/components/light/wink.py
@@ -44,10 +44,10 @@ class WinkLight(WinkToggleDevice):
         brightness = kwargs.get(ATTR_BRIGHTNESS)
 
         if brightness is not None:
-            self.wink.setState(True, brightness / 255)
+            self.wink.set_state(True, brightness / 255)
 
         else:
-            self.wink.setState(True)
+            self.wink.set_state(True)
 
     @property
     def state_attributes(self):
